Add support to configure basic Host redirection
This project adds a custom `nginx.conf` that  supports redirection for
a set
of given domains. The domains should be in the `Host` or
X-Forwarded-Host`
(for cloudfront) of the request header.

This is useful for redirecting from an old domain or add a `www` prefix,
like `cloud.service.gov.uk` to `www.cloud.service.gov.uk`.

To change configure it, add the environment variable `REDIRECT_RULES` as
follows:

    REDIRECT_RULES=cloud.service.gov.uk:www.cloud.service.gov.uk,government-paas-developer-docs.readthedocs.io:www.cloud.service.gov.uk