Add CONTRIBUTING.md
Summarising a conversation with Ben Henley, Oliver Byford, Graham Bleach and
Hector Rivas in #product-page-format on Slack. A link to this page will
appear whenever you open a new pull request.

I have given the team `team-government-paas-people` write access to the
repo, so that they can push branches and merge deployment related changes,
on the understanding that we will always use pull requests.