Prepare the release scripts
We'd like to separate the release scripts, with the thought of docker
using them to run specific tasks.

The build script, is suppose to run the dependency installation with
bower and middle man preparation. It will also copy the nginx file to
the desired location.

The push script, will attempt to blue-green deploy the new application.

The test script has been omitted for the time being.