Explicitly warn that the app lacks CSRF protection
Configuring Cross-Site Request Forgery (CSRF) protection in the app
would require enabling sessions, enabling new Rack::Protection
modules, and adding hidden form fields with secret tokens to our
forms.

At present, and likely long-term, the app does not perform any
authentication of its users. Anyone with internet access can submit
a signup or contact request with any `.gov.uk` email address. As such
it isn't clear that CSRF protection, to prevent your browser secretly
submitting the form from a malicious page, delivers any benefit.

Considering it was unclear how to configure CSRF protection, we want
to ensure this won't be missed if we do one day extend this app to
need it. Sessions are disabled by default, and we've added an explicit
disable of them. We've then labeled it with an explicit warning that
the app lacks CSRF protection. This should make us aware in future.