Restrict view names to whitelisted characters
We're applying slightly 'paranoid' security measures in order to
quickly redeploy a cut-down form of the application. This commit
whitelists the names of view files being loaded to alphanumerics,
dashes and underscores. We can be quite sure that these characters
are insufficient to exploit directory traversal or much of anything
else.