Anonymise user IP and enable beacon features in GA
Set the GA tracker to anonymise user IP addresses, disable display
features and use the beacon transport mode.

Anonymise user IP addresses for all hits sent from the tracker:
https://developers.google.com/analytics/devguides/collection/analyticsjs/ip-anonymization

Disable all display features functionality:
https://developers.google.com/analytics/devguides/collection/analyticsjs/display-features#disabling-display-features

To avoid missing hits in GA when a document unloads, set the tracker to
use the Navigator.sendBeacon() transport mechanism when available. This
will transmit hits to GA asynchronously when the User Agent has an
opportunity to do so, without delaying the unload or affecting the
performance of the next navigation. GA will fallback to XMLHttpRequest
or Image transport mechanisms when navigator.sendBeacon isn't supported
by the user's browser.

https://developers.google.com/analytics/devguides/collection/analyticsjs/sending-hits#specifying_different_transport_mechanisms

The "beacon" feature above requires a websocket connection to the google
analytics endpoint, and so the Content-Security-Policy was updated to
allow this.