Fix browser back button functionality
Currently, when you start filling in the form and on a step > 2 you
decide you'd like to go back and amend your data you need to use the `<
Back` link. The browser back button will kick you back to the getting
started page causing loss of data and the need of starting over.

We're adding state management with the History API in order to mitigate
page change, which should prevent form reseting.