Auto-select user invite section
If the user making a request is not going to be a manager, we'd like to
automatically trigger the users invite section.

In order to make it easier for ourselves, we decided to make this
functionality in JS, meaning it won't be triggered when the page is
rendered with javascript turned off.