Add /migrate-live-services
Taken from the Google Doc written by Paola, including the new sidebar for
both pages in that section. We decided it was better to keep the URL of
`/get-started` even though it doesn't match the "Evaluate the platform" text
for the moment.

The `<title>` has been taken from the H1. We decided together not to use
"Get started with GOV.UK PaaS" as the H1 because it made less sense for this
page. URL for the new page was agreed with Jess and Paola.