Create a Start using GOV.UK PaaS page
We think, that the default page the UAA redirects our tenants to is less
useful for them to get started. We'd like to provide them with as much
information as possible for them to find their way around PaaS.

Stephen has provided us with both the design and the prototype page,
which we decided could be integrated into the product page for being
static.