Relax ruby version constraint.
This was pining the exact ruby version. This resulted in a deployment
failure when the buildpack was updated and dropped support for 2.4.1

This therefore updates the version constraints to use pessimestic
version pinning so that it uses the latest 2.4.x supported by the
buildpack.

Specifying '2.4' in the .travis.yml causes Travis to pick the latest
2.4.x version it has available[1].

It's not possible to specify anything other than an exact ruby version
in the .ruby-version file, so this has been udpated to 2.4.3, which is
the latest 2.4 version at present. I've also stopped reading this in the
Gemfile as it's no longer the source of truth (and it was breaking
Travis which only has 2.4.1 available)

[1]https://docs.travis-ci.com/user/languages/ruby/#Specifying-Ruby-versions-and-implementations