Create contactus/signup tickets in Zendesk
We are planing to move from Deskpro to Zendesk and we want
the product page to create tickets in Zendesk.

In this commit we use the official Zendesk API ruby client[1].

We can use this library directly, no need for a model. The validation
will happen in each corresponding Form object.

We allow use one backend or other by checking the environment variable
`USE_ZENDESK=true`. We can later decommission the Deskpro code as
needed.

We test both backends using rspec.shared_examples.

We additionally move the initialisation of the `ENV[]` variables to
the main `spec_helper.rb` file. Currently they are in the root
scope for each file and change global variable, so they are
actually executed multiple times when each file is loaded.

[1] https://github.com/zendesk/zendesk_api_client_rb